Updated prefabs to all share a base prefab class so every prefab always consist of an entity's 'shapeParams' and 'matterOptions' objects by default, making it easier to instantiate multiple instances of the same prefab config, but still be able to modify each instance's config.  Updated PsyanimScene to be responsible for instantiating prefabs and returning the corresponding entity.
Made the PsyanimPlayfightAgentPrefab and PsyanimArriveAgentPrefabs both a part of PsyanimCore, fixing them up to work in the new prefab system design, and now anyone can create playfight agents or arrive agents with just a few lines of code.

Created a PsyanimClickToMovePlayerPrefab, as well as a class to encapsulate a reusable 'test environment' with a navigation grid called 'TestGridWithObstacles', and updated ClickToMoveTest to use these now, reducing the lines of code in the scene create() method by almost a factor of 10!